@@ -29,6 +29,8 @@ import javax.ws.rs.core.Response.Status;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Entities.EscapeMode;
 import org.jsoup.safety.Whitelist;
 
 import fi.otavanopisto.muikku.model.users.UserEntity;
@@ -465,10 +467,13 @@ public class WorkspaceForumRESTService extends PluginRESTService {
     }
 
     if (sessionController.hasWorkspacePermission(ForumResourcePermissionCollection.FORUM_WRITE_WORKSPACE_MESSAGES, workspaceEntity)) {
+      Document message = Jsoup.parse(Jsoup.clean(newThread.getMessage(), Whitelist.relaxed().addAttributes("a", "target")));
+      message.outputSettings().escapeMode(EscapeMode.xhtml);
+      message.select("a[target]").attr("rel", "noopener noreferer");
       ForumThread thread = forumController.createForumThread(
           forumArea, 
           newThread.getTitle(),
-          Jsoup.clean(newThread.getMessage(), Whitelist.relaxed()), 
+          message.body().toString(), 
           newThread.getSticky(), 
           newThread.getLocked());
   
