@@ -65,7 +65,9 @@
         var itemsContainer = this.element.find('.muikku-sorter-items-container');
         if (readonly) {
           this.element.attr('data-disabled', 'true');
-          $(itemsContainer).sortable("destroy");
+          if ($(itemsContainer).sortable()) {
+            $(itemsContainer).sortable("destroy");
+          }
         } else {
           $(itemsContainer).sortable({
             tolerance: 'pointer',
@@ -110,10 +112,11 @@
         var correctItems = this.options.meta.items;
         var itemsContainer = $(this.element).find('.muikku-sorter-items-container')[0];
         for (var i = 0; i < correctItems.length; i++) {
-          var correctId = correctItems[i].id;
           var userId = userAnswer[i];
+          var userText = this._getItemText(userId); 
+          var correctText = correctItems[i].name;
           var itemElement = $(this.element).find('.muikku-sorter-item[data-item-id="' + userId + '"]')[0];
-          if (correctId == userId) {
+          if (correctText == userText) {
             result.correctAnswers++;
             $(itemElement).addClass('muikku-field-correct-answer');
           }
@@ -127,12 +130,13 @@
           exampleDetails.append( 
             $('<span>').addClass('muikku-field-examples-title').text(getLocaleText('plugin.workspace.assigment.checkAnswers.correctSummary.title'))
           );
+          var capitalize = this.options.meta.capitalize;
           var orientation = this.options.meta.orientation;
           var correctString = '';
           var delimiter = orientation == 'horizontal' ? ' ' : ', ';
           for (var i = 0; i < correctItems.length; i++) {
             var value = correctItems[i].name;
-            if (i == 0 && orientation == 'horizontal') {
+            if (i == 0 && capitalize == true) {
               value = value.charAt(0).toUpperCase() + value.slice(1);
             }
             correctString += i == 0 ? value : delimiter + value; 
@@ -144,6 +148,15 @@
       },
       canCheckAnswer: function() {
         return true;
+      },
+      _getItemText: function(itemId) {
+        var items = this.options.meta.items;
+        for (var i = 0; i < items.length; i++) {
+          if (items[i].id == itemId) {
+            return items[i].name;
+          }
+        }
+        return null;
       },
       _onBeforeAssignmentSubmit: function (event, data) {
         if (data.state == 'UNANSWERED') {
