@@ -40,8 +40,15 @@ $(document).ready(function(){
                 err);
           } else {
             renderDustTemplate('announcer/announcer_widget_items.dust', result, $.proxy(function (text) {
-                  cont.html(text);
-                  cont.perfectScrollbar({"suppressScrollY" : true});
+              cont.html(text);
+              var is_xs = $(window).width() < 769;
+              if (!is_xs) {
+                cont.perfectScrollbar({
+                  wheelSpeed:3,
+                  swipePropagation:false,
+                  wheelPropagation:true
+                });
+              }
             }, this));
           }
         });
